@@ -2,18 +2,18 @@
   "name": "early_access_nft",
   "version": "0.1.0",
   "private": true,
+  "homepage": "./",
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.4",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "@thirdweb-dev/react": "^2.0.0",
-    "@thirdweb-dev/sdk": "^2.0.4",
+    "@thirdweb-dev/react": "^3.6.8",
+    "@thirdweb-dev/sdk": "^3.6.8",
     "@types/jest": "^27.4.1",
-    "@types/react-dom": "^17.0.13",
     "dotenv": "^16.0.0",
     "ethers": "^5.6.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-scripts": "5.0.0",
     "web-vitals": "^2.1.4"
   },
@@ -21,7 +21,8 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "deploy": "npm run build && npx thirdweb upload build/"
   },
   "eslintConfig": {
     "extends": [
@@ -42,6 +43,7 @@
     ]
   },
   "devDependencies": {
+    "@types/react-dom": "^18.0.9",
     "ts-node": "^10.7.0",
     "tslib": "^2.3.1",
     "typescript": "^4.6.2"
