@@ -1,9 +1,11 @@
-import { ThirdwebProvider, ChainId } from '@thirdweb-dev/react';
-import Nft from './components/Nft';
+import { ThirdwebProvider, ChainId } from "@thirdweb-dev/react";
+import Nft from "./components/Nft";
+
+const activeChainId = ChainId.Mumbai;
 
 function App() {
   return (
-    <ThirdwebProvider desiredChainId={ChainId.Rinkeby}>
+    <ThirdwebProvider desiredChainId={activeChainId}>
       <Nft />
     </ThirdwebProvider>
   );
